services/ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
openai==1.3.0
python-dotenv==1.0.0
httpx==0.25.0
tiktoken==0.5.1
transformers==4.35.0
torch==2.1.0
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-multipart==0.0.6
requests==2.31.0
nltk==3.8.1
beautifulsoup4==4.12.2
markdown==3.5.1